Add template for testbed stylesheets
The processing script will convert either the source GDB or
shapefiles (should it support both?) so that the embedded OSM tags
are extracted to their own columns on the data. GeoSolutions has
done this with their implementation by converting to GeoJSON and
then to shapefiles.

The processing script still needs to be written.

The stylesheet is based on the natural earth data stylesheet,
except the centre is on Daraa, Syria.